Add comprehensive SIMD tests for multi-backend support
This commit introduces a new test suite for SIMD operations across various backends including CPU, CUDA, Vulkan, and OneAPI. The tests cover all functions in the `tenzor/backends/cpu/simd.hpp` file, ensuring extensive edge case testing for arithmetic operations, unary functions, activation functions, and performance validation. Key features include:

- Randomized input generation for robustness.
- Approximate equality checks to handle floating-point precision issues.
- Tests for special values and edge cases.
- Performance benchmarks for large array operations.

All tests are instantiated for each backend to ensure consistent behavior across different hardware implementations.

+++ include/tenzor/data/transforms.hpp
@@ -56,7 +56,7 @@ public:
 
         // Assume last dimension is channels
         size_t num_channels = mean_.size();
-        if (shape.back() != num_channels && num_channels > 1) {
+        if (static_cast<size_t>(shape.back()) != num_channels && num_channels > 1) {
             throw std::invalid_argument("Input channels must match normalization parameters");
         }
